change admission_tests.maximum_candidates and admission_test_has_candidate.seat_number from tiny int to small int

+++ database/migrations/2025_01_22_232236_create_admission_tests_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->dateTime('expect_end_at')->nullable();
             $table->unsignedBigInteger('location_id')->nullable();
             $table->unsignedBigInteger('address_id')->nullable();
-            $table->unsignedTinyInteger('maximum_candidates')->nullable();
+            $table->unsignedSmallInteger('maximum_candidates')->nullable();
             $table->boolean('is_free')->default(false);
             $table->boolean('is_public')->default(false);
             $table->timestamps();
